Rename root to cmd
